refactor button and quick list tests
